--- colors.py
@@ -1,31 +1,0 @@
-"""
-This is where all the colors available will be stored. The format is as follows:
-
-'nameofcolor': (  R,   G,   B,   A)
-
-If a component is less than three digits, use spaces for consistency
-"""
-
-
-colors = {'clear': (  0,   0,   0,   0),
-          'fullmenu': (  5,   0,  30,  150),
-          'white': (255, 255, 255),
-          'starwhite': (230, 230, 230),
-          'gray': (120, 120, 120),
-          'darkgray': ( 50,  50,  50),
-          'space': (  0,   0,   7),
-          'black': (  0,   0,   0),
-          'red': (255,   0,   0),
-          'orange': (220, 100,   5),
-          'yellow': (255, 255,   0),
-          'green': (  0, 255,   0),
-          'lime': (120, 255,   0),
-          'mint': (  0, 255, 170),
-          'cyan': (  0, 255, 255),
-          'skyblue': (  0, 110, 255),
-          'blue': (  0,   0, 255),
-          'deepblue': (  5,   0, 20),
-          'indigo': ( 50,   0, 255),
-          'purple': (128,   0, 255),
-          'pink': (255,   0, 255),
-          'rose': (255,   0,  80)}
